requirements.txt
motor==3.3.1
passlib==1.7.4
python-jose==3.5.0
bcrypt==4.3.0
pydantic==2.11.7
python-dotenv==1.1.1
pymongo==4.5.0